Made asset importer menu with support for .obj and .wav files and a way to store them and attach each other.
Made asset importer menu with support for .obj and .wav files and a way to store them and attach each other.
TODO - Make pallet for putting objects in world. Make audio GUI for sounds not attached to WO's. Make a file export tool. Allow movement, rotation, and scaling of WO's and move attached sounds with them.